Problem
A man must travel for given number of days by bus. He can buy either: 1-day ticket for Rs 2 (valid for 1 day) 7 days ticket for Rs 7 (valid for 7 consecutive days) 30 days ticket for Rs 25 (valid for 30 consecutive days) We are given an array; whose elements are the day numbers on which he travels.

For example, if the given array is {1,3,5,8,9,10} then he travels on these 6 days. Now we must minimize the cost of his travel. What would be the best approach to solve this problem?